First attempt at calculating a neutral density tendency
This is the more direct approach to incorporating the minimisation of
neutral density curvature into a vertical grid generation
scheme. Here, we calculate the tendency of a given interface,
according to the local neutral density curvature calculated from the
neighbouring cells and the local buoyancy.

The limiting condition to ensure we don't get oscillations of the grid
seems a little tricky to implement, and will probably have to also be
based on the local buoyancy frequency.