Incorporate neutral density into grid diffusivity
The calcluations in the Climatology notebook have been broken up into
functions that constitute the grid diffusivity algorithm. A
diffusivity coefficient is calcluated from the neutral density
curvature, according to the convergence/divergence of adjacent
interfaces within a column. This doesn't give the most satisfactory
results, due to the strongly discontinuous nature of this coefficient
calculation.