Fix uniform thickness calculation
I was calculating "uniform" thickness by scaling the range [0,1]. That
doesn't work at all! That's fixed, so the initial coordinate from
which we optimise makes more sense now. There's still an issue with
"striping" under neutral density optimisation, which is probably
because we're not doing any displacement limiting based on information
from neigbouring cells. I don't see an obvious solution here because
in the initial state when all interfaces are horizontal, it's easy to
implement a limiter than prevents any deviation from this state!