Add notebook to calculate the 75-level HyCOM1 grid
Here I'm using some of the configuration from OM4_025, which includes
the target densities, the nominal interface positions, the maximum
interface positions, the maximum layer thickness, and using the P1M_H2
regridding interpolation scheme (which luckily is easy to
implement). I'm not doing any Fortran wrapping here, because the
regridding module is a bit too embedded within the rest of the module
to make that an easy task for the moment.