Add salinity-modified HyCOM algorithm
We want the transition region in the high-latitude northern hemisphere
to be slightly shallower to allow for overflows, but at the same time
we need the transition region in the high-latitude southern hemisphere
to be fairly deep in order to ensure the halocline is
well-resolved. To account for this, we adjust the transition
depth/pressure by the local salinity, taking advantage of the fact
that the northern hemisphere is ~1 of your favourite salinity units
more saline. Because layers that are too thin at the surface pose
stability issues for the model, we re-adjust the transition interfaces
back to a minimum thickness after the salinity modification.